+++ src/rex.asynctask/src/rex/asynctask/transports/__init__.py
@@ -4,6 +4,7 @@
 
 
 from .base import AsyncTransport
+from .filesys import FileSysAsyncTransport
 from .localmem import LocalMemoryAsyncTransport
 from .pgsql import PostgresAsyncTransport
 from .redis import RedisAsyncTransport
@@ -11,6 +12,7 @@ from .redis import RedisAsyncTransport
 
 __all__ = (
     'AsyncTransport',
+    'FileSysAsyncTransport',
     'LocalMemoryAsyncTransport',
     'PostgresAsyncTransport',
     'RedisAsyncTransport',