+++ src/rex.asynctask/NEWS.rst
@@ -8,8 +8,8 @@ REX.ASYNCTASK Change Log
 0.5.0 (2017-xx-xx)
 ==================
 
-* Added a basic task scheduling system that allows workers to be triggered via
-  a crontab-like configuration.
+* Added a basic task scheduling system that allows workers and rex.ctl Tasks to
+  be triggered via a crontab-like configuration.
 
 
 0.4.0 (2017-01-19)